BIts of cleanup
