--- addons/pror/CfgAmmo.hpp
@@ -1,10 +1,0 @@
-class CfgAmmo {
-    class B_127x99_Ball;
-    class B_127x99_Ball_Mk211: B_127x99_Ball {
-        explosive = 0.2;
-        indirectHit = 10;
-        indirectHitRange = 0.3;
-        CraterEffects = "ExploAmmoCrater";
-        explosionEffects = "ExploAmmoExplosion";
-    };
-};
