--- addons/pror/CfgEventHandlers.hpp
@@ -1,103 +1,0 @@
-
-class Extended_InitPost_EventHandlers {
-    class Inegal_LandRover_Unarmed {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_7('Inegal_C','Inegal_F','Inegal_F','Inegal_MG','Inegal_ATE','Inegal_Mark','Inegal_ATJ')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_LandRover_FDF_Unarmed {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_7('Inegal_FDF_V','Inegal_FDF_S','Inegal_FDF_S','Inegal_FDF_B','Inegal_FDF_Sh','Inegal_FDF_A','Inegal_FDF_Se')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_VAB_Unarmed {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_2('Inegal_C','Inegal_F')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_VAB_D_Unarmed {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_2('Inegal_Diablerie_C','Inegal_Diablerie_S')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_VAB_HMG {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_1('Inegal_C')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_VAB_D_HMG {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_1('Inegal_Diablerie_C')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_VAB_GMG {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_1('Inegal_C')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_VAB_D_GMG {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_1('Inegal_Diablerie_C')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_TRM_Covered {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_17('Inegal_C','Inegal_F','Inegal_F','Inegal_F','Inegal_F','Inegal_MED','Inegal_MG','Inegal_MG','Inegal_MG','Inegal_Mark','Inegal_Mark','Inegal_Mark','Inegal_AA','Inegal_ATE','Inegal_ATJ','Inegal_ATJ','Inegal_FAC')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_TRM_D_Covered {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_17('Inegal_Diablerie_G','Inegal_Diablerie_C','Inegal_Diablerie_S','Inegal_Diablerie_S','Inegal_Diablerie_S','Inegal_Diablerie_Su','Inegal_Diablerie_B','Inegal_Diablerie_B','Inegal_Diablerie_B','Inegal_Diablerie_Sh','Inegal_Diablerie_Sh','Inegal_Diablerie_Sh','Inegal_Diablerie_A','Inegal_Diablerie_A','Inegal_Diablerie_A','Inegal_Diablerie_Se','Inegal_Diablerie_Se')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_TRM_R_Covered {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_17('Inegal_Recrues_L','Inegal_Recrues_L','Inegal_Recrues_L','Inegal_Recrues_R','Inegal_Recrues_R','Inegal_Recrues_R','Inegal_Recrues_R','Inegal_Recrues_R','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_TRM_Open {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_17('Inegal_C','Inegal_F','Inegal_F','Inegal_F','Inegal_F','Inegal_MED','Inegal_MG','Inegal_MG','Inegal_MG','Inegal_Mark','Inegal_Mark','Inegal_Mark','Inegal_AA','Inegal_ATE','Inegal_ATJ','Inegal_ATJ','Inegal_FAC')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_TRM_D_Open {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_17('Inegal_Diablerie_G','Inegal_Diablerie_C','Inegal_Diablerie_S','Inegal_Diablerie_S','Inegal_Diablerie_S','Inegal_Diablerie_Su','Inegal_Diablerie_B','Inegal_Diablerie_B','Inegal_Diablerie_B','Inegal_Diablerie_Sh','Inegal_Diablerie_Sh','Inegal_Diablerie_Sh','Inegal_Diablerie_A','Inegal_Diablerie_A','Inegal_Diablerie_A','Inegal_Diablerie_Se','Inegal_Diablerie_Se')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_TRM_R_Open {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_17('Inegal_Recrues_L','Inegal_Recrues_L','Inegal_Recrues_L','Inegal_Recrues_R','Inegal_Recrues_R','Inegal_Recrues_R','Inegal_Recrues_R','Inegal_Recrues_R','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C','Inegal_Recrues_C')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_APC_VBCI {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_8('Inegal_C','Inegal_F','Inegal_F','Inegal_MED','Inegal_MG','Inegal_Mark','Inegal_ATE','Inegal_FAC')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_Heli_AW101 {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_16('Inegal_C','Inegal_F','Inegal_F','Inegal_F','Inegal_F','Inegal_MED','Inegal_MG','Inegal_MG','Inegal_MG','Inegal_Mark','Inegal_Mark','Inegal_Mark','Inegal_AA','Inegal_ATE','Inegal_ATJ','Inegal_FAC')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_Heli_MH6 {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_6('Inegal_C','Inegal_F','Inegal_F','Inegal_MG','Inegal_Mark','Inegal_FAC')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_Heli_FDF_MH6 {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_6('Inegal_FDF_V','Inegal_FDF_S','Inegal_FDF_B','Inegal_FDF_Sh','Inegal_FDF_A','Inegal_FDF_Se')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_Heli_D_MH6 {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,[ARR_6('Inegal_Diablerie_C','Inegal_Diablerie_S','Inegal_Diablerie_B','Inegal_Diablerie_Sh','Inegal_Diablerie_A','Inegal_Diablerie_Se')])] call EFUNC(common,populate));
-        };
-    };
-    class Inegal_Diablerie_S {
-        class ADDON {
-            init = QUOTE([ARR_2(_this select 0,1)] call EFUNC(common,setSkill));
-        };
-    };
-};
