--- addons/pror/CfgMagazines.hpp
@@ -1,8 +1,0 @@
-class CfgMagazines {
-    class CUP_10Rnd_127x99_M107;
-    class CUP_10Rnd_127x99_M107_Mk211: CUP_10Rnd_127x99_M107 {
-        displayName = "12.7mm 10Rnd Mk211";
-        ammo = "B_127x99_Ball_Mk211";
-        descriptionShort = "Caliber: 12.7x99mm Mk211 NATO <br/>Rounds: 10 <br/>Used in: M107";
-    };
-};
