+++ addons/pror/units/CfgParas.hpp
@@ -1,0 +1,243 @@
+class PROR_Para : PROR_R {
+    displayName = "Paratrooper";
+    uniformClass = "CFP_U_FieldUniform_M81light";
+    backpack = "B_Parachute";
+    weapons[] = {
+        "arifle_CTAR_blk_F_aco_grn",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "arifle_CTAR_blk_F_aco_grn",
+        "Throw",
+        "Put"
+    };
+    magazines[] = {
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_Tracer_F",
+        "30Rnd_580x42_Mag_Tracer_F",
+        "SmokeShell",
+        "SmokeShell",
+        "HandGrenade"
+    };
+    respawnMagazines[] = {
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_Tracer_F",
+        "30Rnd_580x42_Mag_Tracer_F",
+        "SmokeShell",
+        "SmokeShell",
+        "HandGrenade"
+    };
+};
+class PROR_Para_SL : PROR_R {
+    displayName = "Para Section Leader";
+    uniformClass = "CFP_U_FieldUniform_M81light";
+    backpack = "B_Parachute";
+    weapons[] = {
+        "arifle_CTAR_GL_blk_F_aco_grn",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "arifle_CTAR_GL_blk_F_aco_grn",
+        "Throw",
+        "Put"
+    };
+    magazines[] = {
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_Tracer_F",
+        "30Rnd_580x42_Mag_Tracer_F",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "Smokeshell",
+        "Smokeshell",
+        "Smokeshell"
+    };
+    respawnMagazines[] = {
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_F",
+        "30Rnd_580x42_Mag_Tracer_F",
+        "30Rnd_580x42_Mag_Tracer_F",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "1Rnd_HE_Grenade_shell",
+        "Smokeshell",
+        "Smokeshell",
+        "Smokeshell"
+    };
+    linkedItems[] = {
+        "CFP_Basic_Helmet_M81",
+        "CUP_V_B_RRV_TL",
+        "CFP_ANPVS15_Camo_V3",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "CFP_Shemagh_Neck_M81"
+    };
+    respawnLinkedItems[] = {
+        "CFP_Basic_Helmet_M81",
+        "CUP_V_B_RRV_TL",
+        "CFP_ANPVS15_Camo_V3",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "CFP_Shemagh_Neck_M81"
+    };
+};
+class PROR_Para_AR : PROR_R {
+    displayName = "Para Autorifleman";
+    uniformClass = "CFP_U_FieldUniform_M81light";
+    backpack = "B_Parachute";
+    weapons[] = {
+        "arifle_CTARS_blk_F_aimpoint",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "arifle_CTARS_blk_F_aimpoint",
+        "Throw",
+        "Put"
+    };
+    magazines[] = {
+        "100Rnd_580x42_Mag_F",
+        "100Rnd_580x42_Mag_F",
+        "100Rnd_580x42_Mag_Tracer_F",
+        "100Rnd_580x42_Mag_Tracer_F",
+        "SmokeShell",
+        "SmokeShell",
+        "HandGrenade"
+    };
+    respawnMagazines[] = {
+        "100Rnd_580x42_Mag_F",
+        "100Rnd_580x42_Mag_F",
+        "100Rnd_580x42_Mag_Tracer_F",
+        "100Rnd_580x42_Mag_Tracer_F",
+        "SmokeShell",
+        "SmokeShell",
+        "HandGrenade"
+    };
+    linkedItems[] = {
+        "CFP_Basic_Helmet_M81",
+        "CUP_V_B_RRV_Scout",
+        "CFP_ANPVS15_Camo_V3",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "CFP_Shemagh_Neck_M81"
+    };
+    respawnLinkedItems[] = {
+        "CFP_Basic_Helmet_M81",
+        "CUP_V_B_RRV_Scout",
+        "CFP_ANPVS15_Camo_V3",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "CFP_Shemagh_Neck_M81"
+    };
+};
+class PROR_Para_Marks : PROR_R {
+    displayName = "Para Marksman";
+    uniformClass = "CFP_U_FieldUniform_M81light";
+    backpack = "B_Parachute";
+    weapons[] = {
+        "srifle_DMR_06_olive_F_khs",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "srifle_DMR_06_olive_F_khs",
+        "Throw",
+        "Put"
+    };
+    magazines[] = {
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "SmokeShell",
+        "SmokeShell",
+        "HandGrenade"
+    };
+    respawnMagazines[] = {
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "20Rnd_762x51_Mag",
+        "SmokeShell",
+        "SmokeShell",
+        "HandGrenade"
+    };
+};
+class PROR_Para_AT : PROR_Para {
+    displayName = "Para AT";
+    uniformClass = "CFP_U_FieldUniform_M81light";
+    backpack = "B_Parachute";
+    weapons[] = {
+        "arifle_CTAR_blk_F_aco_grn",
+        "CUP_launch_M72A6",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "arifle_CTAR_blk_F_aco_grn",
+        "CUP_launch_M72A6",
+        "Throw",
+        "Put"
+    };
+};
