+++ addons/pror/units/CfgSF.hpp
@@ -1,0 +1,290 @@
+class PROR_SF_R : PROR_Soldier_Base {
+    scope = 2;
+    displayName = "SF Rifleman";
+    editorSubcategory = "EdSubcat_Personnel";
+    uniformClass = "CFP_GUER_M81";
+    model = "\A3\characters_F_gamma\Guerrilla\ig_guerrilla2_1.p3d";
+    backpack = "PROR_SF_AssPack";
+    weapons[] = {
+        "SMA_MK18MOEOD_CCO_LLCTop_blk",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "SMA_MK18MOEOD_CCO_LLCTop_blk",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+    magazines[] = {
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "CUP_30Rnd_556x45_Emag_Tracer_Yellow",
+        "CUP_30Rnd_556x45_Emag_Tracer_Yellow",
+        "CUP_30Rnd_556x45_Emag_Tracer_Yellow",
+        "CUP_30Rnd_556x45_Emag_Tracer_Yellow",
+        "SmokeShell",
+        "SmokeShell",
+        "SmokeShell",
+        "HandGrenade",
+        "HandGrenade"
+    };
+    respawnMagazines[] = {
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "CUP_30Rnd_556x45_Emag_Tracer_Yellow",
+        "CUP_30Rnd_556x45_Emag_Tracer_Yellow",
+        "CUP_30Rnd_556x45_Emag_Tracer_Yellow",
+        "CUP_30Rnd_556x45_Emag_Tracer_Yellow",
+        "SmokeShell",
+        "SmokeShell",
+        "SmokeShell",
+        "HandGrenade",
+        "HandGrenade"
+    };
+    items[] = {
+        "ACE_EarPlugs",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_morphine",
+        "ACE_morphine",
+        "ACE_morphine",
+        "ACE_tourniquet",
+        "ACE_tourniquet"
+    };
+    respawnItems[] = {
+        "ACE_EarPlugs",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_morphine",
+        "ACE_morphine",
+        "ACE_morphine",
+        "ACE_tourniquet",
+        "ACE_tourniquet"
+    };
+    linkedItems[] = {
+        "CFP_OPS2017_Helmet_M81",
+        "CFP_PlateCarrier1_M81",
+        "NVGogglesB_blk_F",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "G_Bandanna_oli"
+    };
+    respawnLinkedItems[] = {
+        "CFP_OPS2017_Helmet_M81",
+        "CFP_PlateCarrier1_M81",
+        "NVGogglesB_blk_F",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "G_Bandanna_oli"
+    };
+};
+class PROR_SF_TL : PROR_SF_R {
+    scope = 2;
+    displayName = "SF Team Leader";
+    editorSubcategory = "EdSubcat_Personnel";
+    uniformClass = "CFP_GUER_M81";
+    model = "\A3\characters_F_gamma\Guerrilla\ig_guerrilla2_1.p3d";
+    backpack = "PROR_SF_TL_AssPack";
+    weapons[] = {
+        "SMA_MK18OD_GL_CCO_LLCTop_blk",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "SMA_MK18OD_GL_CCO_LLCTop_blk",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+};
+class PROR_SF_Med : PROR_SF_R {
+    scope = 2;
+    displayName = "SF Medic";
+    editorSubcategory = "EdSubcat_Personnel";
+    uniformClass = "CFP_GUER_M81";
+    model = "\A3\characters_F_gamma\Guerrilla\ig_guerrilla2_1.p3d";
+    backpack = "PROR_SF_MedicBag";
+    weapons[] = {
+        "SMA_MK18MOEOD_CCO_LLCTop_blk",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "SMA_MK18MOEOD_CCO_LLCTop_blk",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+};
+class PROR_SF_Marks : PROR_SF_R {
+    scope = 2;
+    displayName = "SF Marksman";
+    editorSubcategory = "EdSubcat_Personnel";
+    uniformClass = "CFP_GUER_M81";
+    model = "\A3\characters_F_gamma\Guerrilla\ig_guerrilla2_1.p3d";
+    backpack = "";
+    weapons[] = {
+        "SMA_HK417_16in_ACOG",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "SMA_HK417_16in_ACOG",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+    magazines[] = {
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "SMA_20Rnd_762x51mm_M80A1_EPR",
+        "SMA_20Rnd_762x51mm_M80A1_EPR",
+        "SMA_20Rnd_762x51mm_M80A1_EPR",
+        "SMA_20Rnd_762x51mm_M80A1_EPR",
+        "SMA_20Rnd_762x51mm_M80A1_EPR"
+    };
+    respawnMagazines[] = {
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "SMA_20Rnd_762x51mm_M80A1_EPR",
+        "SMA_20Rnd_762x51mm_M80A1_EPR",
+        "SMA_20Rnd_762x51mm_M80A1_EPR",
+        "SMA_20Rnd_762x51mm_M80A1_EPR",
+        "SMA_20Rnd_762x51mm_M80A1_EPR"
+    };
+    linkedItems[] = {
+        "CFP_OPS2017_Helmet_M81",
+        "CFP_PlateCarrier1_M81",
+        "NVGogglesB_blk_F",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "G_Bandanna_oli"
+    };
+    respawnLinkedItems[] = {
+        "CFP_OPS2017_Helmet_M81",
+        "CFP_PlateCarrier1_M81",
+        "NVGogglesB_blk_F",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "G_Bandanna_oli"
+    };
+};
+class PROR_SF_AR : PROR_SF_R {
+    scope = 2;
+    displayName = "SF Autorifleman";
+    editorSubcategory = "EdSubcat_Personnel";
+    uniformClass = "CFP_GUER_M81";
+    model = "\A3\characters_F_gamma\Guerrilla\ig_guerrilla2_1.p3d";
+    backpack = "PROR_SF_Kitbag_AR";
+    weapons[] = {
+        "CUP_lmg_m249_SQuantoon_LDS",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+    respawnWeapons[] = {
+        "CUP_lmg_m249_SQuantoon_LDS",
+        "hgun_ACPC2_F",
+        "Throw",
+        "Put"
+    };
+    magazines[] = {
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "CUP_100Rnd_TE4_Yellow_Tracer_556x45_M249",
+        "SmokeShell",
+        "SmokeShell",
+        "SmokeShell"
+    };
+    respawnMagazines[] = {
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "9Rnd_45ACP_Mag",
+        "CUP_100Rnd_TE4_Yellow_Tracer_556x45_M249",
+        "SmokeShell",
+        "SmokeShell",
+        "SmokeShell"
+    };
+    items[] = {
+        "ACE_EarPlugs",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_morphine",
+        "ACE_morphine",
+        "ACE_morphine",
+        "ACE_tourniquet",
+        "ACE_tourniquet"
+    };
+    respawnItems[] = {
+        "ACE_EarPlugs",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_elasticBandage",
+        "ACE_morphine",
+        "ACE_morphine",
+        "ACE_morphine",
+        "ACE_tourniquet",
+        "ACE_tourniquet"
+    };
+    linkedItems[] = {
+        "CFP_OPS2017_Helmet_M81",
+        "CFP_PlateCarrier1_M81",
+        "NVGogglesB_blk_F",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "G_Bandanna_oli"
+    };
+    respawnLinkedItems[] = {
+        "CFP_OPS2017_Helmet_M81",
+        "CFP_PlateCarrier1_M81",
+        "NVGogglesB_blk_F",
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemRadioAcreFlagged",
+        "ItemGPS",
+        "Binocular",
+        "G_Bandanna_oli"
+    };
+};
