+++ addons/pror/vehicles/CfgJet.hpp
@@ -1,16 +1,11 @@
 // Harrier
 class CUP_B_AV8B_DYN_USMC;
-class PROR_Jet_Harrier: CUP_B_AV8B_DYN_USMC {
+class PROR_Jet_Harrier : CUP_B_AV8B_DYN_USMC {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
-    editorPreview = "CUP\AirVehicles\CUP_AirVehicles_Av8B\Data\preview\CUP_B_AV8B_CAP_USMC.jpg";
     crew = "PROR_Jet_Pilot";
     typicalCargo[] = {
         "PROR_Jet_Pilot"
     };
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
 };
