+++ addons/pror/vehicles/CfgLAV.hpp
@@ -1,18 +1,12 @@
 class CUP_B_LAV25M240_green;
-class PROR_LAV: CUP_B_LAV25M240_green {
+class PROR_LAV : CUP_B_LAV25M240_green {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "LAV (M240)";
-    editorPreview = "CUP\WheeledVehicles\CUP_WheeledVehicles_LAV25\Data\preview\CUP_B_LAV25M240_green.jpg";
     crew = "PROR_Crew";
     typicalCargo[] = {
         "PROR_Crew",
         "PROR_Crew"
     };
-    
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
 };
