+++ addons/pror/vehicles/CfgMRAP.hpp
@@ -4,48 +4,40 @@ class PROR_MRAP_Unarmed : B_T_MRAP_01_F {
     side = 0;
     faction = "UKSF_PROR";
     displayName = "MRAP Unarmed";
-    editorPreview = QPATHTOEF(common,data\previews\PROR_VAB_Unarmed.jpg);
+    // editorPreview = QPATHTOEF(common,data\previews\PROR_VAB_Unarmed.jpg);
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R"
     };
-    hiddenSelectionsTextures[] = { "\A3\soft_F_Exp\MRAP_01\data\MRAP_01_base_olive_CO.paa","\A3\soft_F_Exp\MRAP_01\data\MRAP_01_adds_olive_CO.paa","\A3\Data_F_Exp\Vehicles\Turret_olive_CO.paa" };
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
+    hiddenSelectionsTextures[] = {
+        "\A3\soft_F_Exp\MRAP_01\data\MRAP_01_base_olive_CO.paa",
+        "\A3\soft_F_Exp\MRAP_01\data\MRAP_01_adds_olive_CO.paa",
+        "\A3\Data_F_Exp\Vehicles\Turret_olive_CO.paa"
+    };
 };
 
- class B_T_MRAP_01_hmg_F;
- class PROR_MRAP_HMG: B_T_MRAP_01_hmg_F {
-     scope = 2;
-     side = 0;
-     faction = "UKSF_PROR";
-     displayName = "MRAP HMG";
-     editorPreview = QPATHTOEF(common,data\previews\PROR_VAB_HMG.jpg);
-     crew = "PROR_R";
-     typicalCargo[] = {
+class B_T_MRAP_01_hmg_F;
+class PROR_MRAP_HMG : B_T_MRAP_01_hmg_F {
+    scope = 2;
+    side = 0;
+    faction = "UKSF_PROR";
+    displayName = "MRAP HMG";
+    // editorPreview = QPATHTOEF(common,data\previews\PROR_VAB_HMG.jpg);
+    crew = "PROR_R";
+    typicalCargo[] = {
         "PROR_R"
-     };
-     class TransportMagazines {};
-     class TransportItems {};
-     class TransportWeapons {};
-     class TransportBackpacks {};
- };
- 
- class B_T_MRAP_01_gmg_F;
- class PROR_MRAP_GMG: B_T_MRAP_01_gmg_F {
-     scope = 2;
-     side = 0;
-     faction = "UKSF_PROR";
-     displayName = "MRAP GMG";
-     editorPreview = QPATHTOEF(common,data\previews\PROR_VAB_GMG.jpg);
-     crew = "PROR_R";
-     typicalCargo[] = {
+    };
+};
+
+class B_T_MRAP_01_gmg_F;
+class PROR_MRAP_GMG : B_T_MRAP_01_gmg_F {
+    scope = 2;
+    side = 0;
+    faction = "UKSF_PROR";
+    displayName = "MRAP GMG";
+    // editorPreview = QPATHTOEF(common,data\previews\PROR_VAB_GMG.jpg);
+    crew = "PROR_R";
+    typicalCargo[] = {
         "PROR_R"
-     };
-     class TransportMagazines {};
-     class TransportItems {};
-     class TransportWeapons {};
-     class TransportBackpacks {};
- };
+    };
+};
