--- addons/pror/vehicles/CfgPolaris
@@ -1,16 +1,0 @@
-
- class B_LSV_01_armed_F;
- class PROR_Polaris_HMG : B_LSV_01_armed_F {
-    scope = 2;
-    side = 0;
-    faction = "UKSF_PROR";
-    displayName = "SF Polaris HMG";
-    editorPreview = QPATHTOEF(common,data\previews\PROR_VAB_Unarmed.jpg);
-    crew = "PROR_SF_R";
-    hiddenSelections[] = {"Camo_1","Camo_2","Camo_3","Camo_4"};
-    hiddenSelectionsTextures[] = { "\A3\Soft_F_Exp\LSV_01\Data\NATO_LSV_01_olive_CO.paa","\A3\Soft_F_Exp\LSV_01\Data\NATO_LSV_02_olive_CO.paa","\A3\Soft_F_Exp\LSV_01\Data\NATO_LSV_03_olive_CO.paa","\A3\Soft_F_Exp\LSV_01\Data\NATO_LSV_Adds_olive_CO.paa" };
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
-};
