+++ addons/pror/vehicles/CfgStatic.hpp
@@ -1,11 +1,10 @@
-// m119
+// M119
 class CUP_B_M119_US;
-class PROR_Static_M119: CUP_B_M119_US {
+class PROR_Static_M119 : CUP_B_M119_US {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "M119";
-    editorPreview = "CUP\Creatures\People\Military\CUP_Creatures_StaticWeapons\Data\preview\CUP_B_M119_US.jpg";
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R"
@@ -13,27 +12,25 @@ class PROR_Static_M119: CUP_B_M119_US {
 };
 
 // HMG
-
 class B_HMG_01_high_F;
-class PROR_Static_HMG_High: B_HMG_01_high_F {
+class PROR_Static_HMG_High : B_HMG_01_high_F {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "HMG (High)";
-    editorPreview = QPATHTOEF(common,data\previews\PROR_Static_GMG.jpg);
+    // editorPreview = QPATHTOEF(common,data\previews\PROR_Static_GMG.jpg);
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R"
     };
 };
-
 class B_HMG_01_F;
-class PROR_Static_HMG_Low: B_HMG_01_F {
+class PROR_Static_HMG_Low : B_HMG_01_F {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "HMG (Low)";
-    editorPreview = QPATHTOEF(common,data\previews\PROR_Static_GMG.jpg);
+    // editorPreview = QPATHTOEF(common,data\previews\PROR_Static_GMG.jpg);
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R"
@@ -42,25 +39,24 @@ class PROR_Static_HMG_Low: B_HMG_01_F {
 
 // GMG
 class B_GMG_01_F;
-class PROR_Static_GMG_Low: B_GMG_01_F {
+class PROR_Static_GMG_Low : B_GMG_01_F {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "GMG (Low)";
-    editorPreview = QPATHTOEF(common,data\previews\PROR_Static_GMG.jpg);
+    // editorPreview = QPATHTOEF(common,data\previews\PROR_Static_GMG.jpg);
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R"
     };
 };
-
 class B_GMG_01_high_F;
-class PROR_Static_GMG_High: B_GMG_01_high_F {
+class PROR_Static_GMG_High : B_GMG_01_high_F {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "GMG (High)";
-    editorPreview = QPATHTOEF(common,data\previews\PROR_Static_GMG.jpg);
+    // editorPreview = QPATHTOEF(common,data\previews\PROR_Static_GMG.jpg);
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R"
@@ -69,12 +65,12 @@ class PROR_Static_GMG_High: B_GMG_01_high_F {
 
 // Mortar
 class B_G_Mortar_01_F;
-class PROR_Static_Mortar: B_G_Mortar_01_F {
+class PROR_Static_Mortar : B_G_Mortar_01_F {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "Mk6";
-    editorPreview = QPATHTOEF(common,data\previews\PROR_Static_Mortar.jpg);
+    // editorPreview = QPATHTOEF(common,data\previews\PROR_Static_Mortar.jpg);
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R"
@@ -82,17 +78,14 @@ class PROR_Static_Mortar: B_G_Mortar_01_F {
 };
 
 // AA
-
 class B_T_Static_AA_F;
-class PROR_Static_AA: B_T_Static_AA_F {
+class PROR_Static_AA : B_T_Static_AA_F {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "AA Launcher";
-    editorPreview = "\A3\EditorPreviews_F_Exp\Data\CfgVehicles\B_T_Static_AA_F.jpg";
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R"
     };
 };
-
