+++ addons/pror/vehicles/CfgStriker.hpp
@@ -1,56 +1,38 @@
 class CUP_B_M1126_ICV_MK19_Woodland;
-class PROR_Striker_GMG: CUP_B_M1126_ICV_MK19_Woodland {
+class PROR_Striker_GMG : CUP_B_M1126_ICV_MK19_Woodland {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "Striker GMG";
-    editorPreview = "CUP\WheeledVehicles\CUP_WheeledVehicles_Stryker\Data\preview\CUP_B_M1126_ICV_MK19_Woodland.jpg";
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R",
         "PROR_SL"
     };
-    
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
 };
 
 class CUP_B_M1126_ICV_M2_Woodland;
-class PROR_Striker_HMG: CUP_B_M1126_ICV_M2_Woodland {
+class PROR_Striker_HMG : CUP_B_M1126_ICV_M2_Woodland {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "Striker HMG";
-    editorPreview = "CUP\WheeledVehicles\CUP_WheeledVehicles_Stryker\Data\preview\CUP_B_M1126_ICV_M2_Woodland.jpg";
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R",
         "PROR_SL"
     };
-    
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
 };
 
 class CUP_B_M1128_MGS_Woodland;
-class PROR_Striker_AT: CUP_B_M1128_MGS_Woodland {
+class PROR_Striker_AT : CUP_B_M1128_MGS_Woodland {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "Striker AT";
-    editorPreview = "CUP\WheeledVehicles\CUP_WheeledVehicles_Stryker\Data\preview\CUP_B_M1128_MGS_Woodland.jpg";
     crew = "PROR_R";
     typicalCargo[] = {
         "PROR_R",
         "PROR_SL"
     };
-    
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
 };
