+++ addons/pror/vehicles/CfgWiesel.hpp
@@ -1,39 +1,37 @@
 class I_LT_01_scout_F;
-class PROR_AA_Radar: I_LT_01_scout_F {
+class PROR_AA_Radar : I_LT_01_scout_F {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "Wiesel AA (RADAR)";
-    editorPreview = "\A3\EditorPreviews_F_Tank\Data\CfgVehicles\I_LT_01_scout_F.jpg";
-    hiddenSelectionsTextures[] = {"A3\armor_f_tank\lt_01\data\lt_01_main_olive_co.paa", "A3\armor_f_tank\lt_01\data\lt_01_at_olive_co.paa", "A3\Armor_F\Data\camonet_AAF_Digi_Green_CO.paa", "A3\armor_f\data\cage_olive_co.paa"};
+    hiddenSelectionsTextures[] = {
+        "A3\armor_f_tank\lt_01\data\lt_01_main_olive_co.paa",
+        "A3\armor_f_tank\lt_01\data\lt_01_at_olive_co.paa",
+        "A3\Armor_F\Data\camonet_AAF_Digi_Green_CO.paa",
+        "A3\armor_f\data\cage_olive_co.paa"
+    };
     crew = "PROR_Crew";
     typicalCargo[] = {
         "PROR_Crew",
         "PROR_Crew"
     };
-    
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
 };
 
 class I_LT_01_AA_F;
-class PROR_AA_Ozelot: I_LT_01_AA_F {
+class PROR_AA_Ozelot : I_LT_01_AA_F {
     scope = 2;
     side = 0;
     faction = "UKSF_PROR";
     displayName = "Wiesel AA (Ozelot)";
-    editorPreview = "\A3\EditorPreviews_F_Tank\Data\CfgVehicles\I_LT_01_AA_F.jpg";
-    hiddenSelectionsTextures[] = {"A3\armor_f_tank\lt_01\data\lt_01_main_olive_co.paa", "A3\armor_f_tank\lt_01\data\lt_01_at_olive_co.paa", "A3\Armor_F\Data\camonet_AAF_Digi_Green_CO.paa", "A3\armor_f\data\cage_olive_co.paa"};
+    hiddenSelectionsTextures[] = {
+        "A3\armor_f_tank\lt_01\data\lt_01_main_olive_co.paa",
+        "A3\armor_f_tank\lt_01\data\lt_01_at_olive_co.paa",
+        "A3\Armor_F\Data\camonet_AAF_Digi_Green_CO.paa",
+        "A3\armor_f\data\cage_olive_co.paa"
+    };
     crew = "PROR_Crew";
     typicalCargo[] = {
         "PROR_Crew",
         "PROR_Crew"
     };
-    
-    class TransportMagazines {};
-    class TransportItems {};
-    class TransportWeapons {};
-    class TransportBackpacks {};
 };
